+++ source/Lucid/Controls/LucidCheckBox.cs
@@ -281,37 +281,37 @@ public class LucidCheckBox : CheckBox
 
         var size = ThemeProvider.Theme.Sizes.CheckBoxSize;
 
-        var textColor = ThemeProvider.Theme.Colors.LightText;
-        var borderColor = ThemeProvider.Theme.Colors.LightText;
-        var fillColor = Checked ? ThemeProvider.Theme.Colors.DarkBackground : ThemeProvider.Theme.Colors.LightestBackground;
+        var textColor = ThemeProvider.Theme.Colors.TextPrimary;
+        var borderColor = ThemeProvider.Theme.Colors.TextPrimary;
+        var fillColor = Checked ? ThemeProvider.Theme.Colors.BackgroundPrimary : ThemeProvider.Theme.Colors.SurfaceHighlight;
 
         if (Enabled)
         {
             if (Focused)
             {
-                borderColor = ThemeProvider.Theme.Colors.ControlHighlight;
-                fillColor = ThemeProvider.Theme.Colors.MainAccent;
+                borderColor = ThemeProvider.Theme.Colors.Accent;
+                fillColor = ThemeProvider.Theme.Colors.Accent;
             }
 
             if (_controlState == LucidControlState.Hover)
             {
-                borderColor = ThemeProvider.Theme.Colors.ControlHighlight;
-                fillColor = ThemeProvider.Theme.Colors.MainAccent;
+                borderColor = ThemeProvider.Theme.Colors.Accent;
+                fillColor = ThemeProvider.Theme.Colors.Accent;
             }
             else if (_controlState == LucidControlState.Pressed)
             {
-                borderColor = ThemeProvider.Theme.Colors.GreyHighlight;
-                fillColor = ThemeProvider.Theme.Colors.GreySelection;
+                borderColor = ThemeProvider.Theme.Colors.SurfaceDefault;
+                fillColor = ThemeProvider.Theme.Colors.SurfaceHighlight;
             }
         }
         else
         {
-            textColor = ThemeProvider.Theme.Colors.DisabledText;
-            borderColor = ThemeProvider.Theme.Colors.GreyHighlight;
-            fillColor = ThemeProvider.Theme.Colors.GreySelection;
+            textColor = ThemeProvider.Theme.Colors.TextDisabled;
+            borderColor = ThemeProvider.Theme.Colors.SurfaceDefault;
+            fillColor = ThemeProvider.Theme.Colors.SurfaceHighlight;
         }
 
-        var backColor = UseBackColorProperty ? this.BackColor : ThemeProvider.Theme.Colors.MainBackgroundColor;
+        var backColor = UseBackColorProperty ? this.BackColor : ThemeProvider.Theme.Colors.BackgroundSecondary;
 
         using (var sg = new Common.SaveableGraphicsState(g))
         {
@@ -332,7 +332,7 @@ public class LucidCheckBox : CheckBox
 
                 if (Checked)
                 {
-                    using (var b = new SolidBrush(ThemeProvider.Theme.Colors.MainAccent))
+                    using (var b = new SolidBrush(ThemeProvider.Theme.Colors.Accent))
                     {
                         // Blue fill color
                         g.FillPath(b, Helper.RoundedRectangleHelper.CreateRoundedRectanglePath(new Rectangle(1, (rect.Height / 2) - (size / 2) + 1, size - 1, size - 1), 2));